securing todo APIs

+++ functions/APIs/todos.js
@@ -3,6 +3,7 @@ const { db } = require('../util/admin');
 exports.getAllTodos = (request, response) => {
     db
         .collection('todos')
+        .where('username', '==', request.user.username)
         .orderBy('createdAt', 'desc')
         .get()
         .then((data) => {
@@ -33,6 +34,7 @@ exports.postOneTodo = (request, response) => {
     }
 
     const newTodoItem = {
+        username: request.user.username,
         title: request.body.title,
         body: request.body.body,
         createdAt: new Date().toISOString()
@@ -58,6 +60,9 @@ exports.deleteTodo = (request, response) => {
         .then((doc) => {
             if (!doc.exists) {
                 return response.status(404).json({ error: 'Todo not found' })
+            }
+            if (doc.data().username !== request.user.username) {
+                return response.status(403).json({error: 'Unauthorized'});
             }
             return document.delete();
         })
